Trigger notifications after GF cache cleared
The products cache doesn't get cleared until priority 10 of `gravityview/edit_entry/after_update`, which is well after when the `gform_after_update_entry` action was fired. This led to notifications (and in my case PDFs) that had out-of-date product details. This change will ensure the notifications aren't fired until after the products cache is cleared.